refactor: limit store 분리
