--- src/features/filter/search/model/store.ts
@@ -1,11 +1,0 @@
-import { create } from "zustand"
-
-interface SearchQueryStore {
-  searchQuery: string
-  setSearchQuery: (searchQuery: string) => void
-}
-
-export const useSearchQueryStore = create<SearchQueryStore>((set) => ({
-  searchQuery: "",
-  setSearchQuery: (searchQuery) => set({ searchQuery }),
-}))
